feat/tech: implement a more appropriate IDB wrapper
